requirements.txt
requests
beautifulsoup4
apscheduler
facebook-sdk
python-dotenv
